+++ mdutils/amber/__init__.py
@@ -1,10 +1,11 @@
 from mdutils.amber.prmtop import Prmtop, PrmtopMeta
 from mdutils.amber.restart import Restart, RestartMeta
 from mdutils.amber.inpcrd import Inpcrd, InpcrdMeta
-from mdutils.amber.groupfile import groupfile
+from mdutils.amber.groupfile import write_groupfile_block, dump_groupfile
 
 __all__ = [
-    "write_groupfile",
+    "write_groupfile_block",
+    "dump_groupfile",
     "Prmtop",
     "PrmtopMeta",
     "Restart",
